end work using FragmentPagerAdapter
