+++ app/src/main/java/ru/firstSet/kotlinOne/Data/Movie.kt
@@ -5,6 +5,7 @@ import android.os.Parcelable
 import ru.firstSet.kotlinOne.Genre
 import kotlinx.parcelize.Parcelize
 import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
 
 @Parcelize
 data class Movie(
@@ -24,3 +25,38 @@ data class Movie(
         var adult: Int,
         var vote_count: Int
 ): Parcelable
+
+@Serializable
+data class MovieForSeach(
+        @SerialName("id")
+        val id: Long,
+        @SerialName("backdrop_path")
+        val backdropPicture: String?,
+        @SerialName("title")
+        val title: String,
+        @SerialName("poster_path")
+        val posterPicture: String?,
+        @SerialName("genre_ids")
+        val genreIds: List<Int>,
+        @SerialName("vote_average")
+        val vote_average: Float,
+        val overview: String,
+        val adult: Boolean,
+        val vote_count: Int
+)
+
+@Serializable
+data class ResultMovie(
+        @SerialName("page")
+        val page: Int,
+        @SerialName("results")
+        val movieForSeach: List<MovieForSeach>
+)
+
+@Serializable
+data class ResultDetails(
+        @SerialName("id")
+        val id: Int,
+        @SerialName("runtime")
+        val runtime: Int
+)
