+++ app/src/main/res/layout/fragment_movies_details.xml
@@ -3,8 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    >
+    android:layout_height="match_parent">
 
     <ScrollView
         android:layout_width="match_parent"
@@ -16,17 +15,18 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             tools:context=".MovieDetailsActivity">
+
             <View
                 android:id="@+id/fmdBackground"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                android:contentDescription="@string/todo"
-                android:scaleType="fitXY"
                 android:layout_alignParentStart="true"
                 android:layout_alignParentTop="true"
                 android:background="@drawable/bg_details"
-                app:layout_constraintEnd_toEndOf="parent"
+                android:contentDescription="@string/todo"
+                android:scaleType="fitXY"
                 app:layout_constraintBottom_toBottomOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toTopOf="parent" />
 
@@ -34,21 +34,20 @@
                 android:id="@+id/fmdPoster"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
+                android:layout_margin="4dp"
+                android:background="@drawable/mask"
                 android:contentDescription="@string/todo"
                 android:scaleType="fitXY"
-                android:layout_margin="4dp"
-                android:src="@drawable/orig"
-                android:background="@drawable/mask"
+                app:layout_constraintBottom_toBottomOf="@id/fmdGuidelineHorizontalCenter"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
-                app:layout_constraintBottom_toBottomOf="@id/fmdGuidelineHorizontalCenter"
-                />
+                app:layout_constraintTop_toTopOf="parent" />
+
             <ProgressBar
                 android:id="@+id/progressBarMovieDetails"
-                android:visibility="invisible"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
+                android:visibility="invisible"
                 tools:ignore="MissingConstraints" />
 
             <androidx.constraintlayout.widget.Guideline
