+++ app/src/main/res/layout/view_holder_movie.xml
@@ -21,11 +21,10 @@
         android:layout_width="@dimen/fmlImageOrigWidth"
         android:layout_height="@dimen/fmlImageOrigHeight"
         android:layout_marginStart="2dp"
+        android:layout_marginTop="3dp"
         android:layout_marginEnd="2dp"
         android:layout_marginBottom="48dp"
         android:contentDescription="@string/todo"
-        android:foreground="@drawable/mask"
-        android:src="@drawable/avengers_end_game"
         app:layout_constraintBottom_toBottomOf="@id/fmlBackGroundMovie"
         app:layout_constraintEnd_toEndOf="@+id/fmlBackGroundMovie"
         app:layout_constraintStart_toStartOf="@+id/fmlBackGroundMovie"
