+++ app/src/main/res/values/colors.xml
@@ -15,4 +15,5 @@
     <color name="colorBackgroud">#191926</color>
     <color name="colorRedText">#FF3466</color>
     <color name="colorTextGrey">#6D6D80</color>
+    <color name="colorLinePager">#8036E7</color>
 </resources>
