+++ app/src/main/res/values/strings.xml
@@ -15,7 +15,7 @@
     <string name="chrisHemsworth">Chris \nHemsworth</string>
     <string name="todo">TODO</string>
     <string name="hello_blank_fragment">Hello blank fragment</string>
-    <string name="fmlWarszawaPo">Movies list</string>
+    <string name="fmlWarszawaPo">Warszawa, Poland</string>
     <string name="fmlSomeId">13+</string>
     <string name="fmlTag">Action, Adventure, Drama</string>
     <string name="fmlTextViewMin">min</string>
