+++ app/src/main/res/values/themes.xml
@@ -1,6 +1,6 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Theme.KotlinOne" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
+    <style name="Theme.KotlinOne" parent="Theme.MaterialComponents.DayNight.NoActionBar">
         <!-- Primary brand color. -->
         <item name="colorPrimary">@color/purple500</item>
         <item name="colorPrimaryVariant">@color/purple700</item>
